move something and delete aggregate.go
