--- executor/aggregate.go
@@ -1,35 +1,0 @@
-// Copyright 2016 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package executor
-
-import (
-	"github.com/pingcap/tidb/util/chunk"
-)
-
-// HashAggInput indicates the input of hash agg exec.
-type HashAggInput struct {
-	chk *chunk.Chunk
-	// giveBackCh is bound with specific partial worker,
-	// it's used to reuse the `chk`,
-	// and tell the data-fetcher which partial worker it should send data to.
-	giveBackCh chan<- *chunk.Chunk
-}
-
-// HashAggIntermData indicates the intermediate data of aggregation execution.
-type HashAggIntermData struct {
-	groupKeys        []string
-	cursor           int
-	partialResultMap aggPartialResultMapper
-}
